Add retrying
This commit makes sure that we retry the request to Jira when it ends
with an error. Sometimes you might hit 429 HTTP error when too many
requests are made. If you hit this error then you lose chunk of Jira
tickets because the script does not retry to obtain them again.